feat: add CI workflows for linting, testing, and building (#11)
* feat: add pre-commit github workflow

* fix: relax python version requirement to >= 3.12

* feat: add ci workflow for testing vision3d library

* fix: do not persist git credentials after checkout

* fix: remove ruff from pre-commit hooks and use via uv run instead

* fix: address ruff errors after relaxing pyhon minimum supported version to 3.12

* feat: add ci workflow for building, linting and testing vision3d

* fix: run all ci workflow jobs inside nvidia/cuda container

* fix: only run on PRs to main

* fix: let uv determine the right python version to install based on .python-version file

* vision3d wheel building

torch requires setuptools<82

* fix: sign compare warnings

* fix: torch matrix pinning during build

* fix: match nvidia/cuda docker image version to torch version in workflow matricies

* fix: set TORCH_CUDA_ARCH_LIST

* fix: guard cuda build behind FORCE_CUDA env var like in torchvision

* fix: add concurrency groups to all workflows

* fix: add numpy build dependency

* fix: make sdist builds work

* fix: free disk space on runners before build

* fix: only delete dotnet installation, otherwise takes too long

* fix: specify version dynamically, including cuda PEP 440 local version identifer

* fix: include version.txt in sdist

* fix: sdist build with custom git version

* fix: docstring at top of setup.py

* fix: delete some more dirs

* fix: use latest minor torch versions

* fix: support python 3.12 via typing_extensions and bump minimum pytorch version to 2.7

* fix: install libs required by opencv transitive dependency

* fix: include cuda tag in test task name

* fix: make typing-extensions dependency optional depending on python version

* fix: add more opencv apt dependencies

* fix: do not persist credentials after checkouts

* fix: make uv cache live on same fs as venv for optimal performance

* fix: make torch 2.8 the minimum supported version as it is the first release to support cu128.

We don't want to support releases that don't support Blackwell gpus

* fix: separate build matrix from test matrix

* fix: bump minimum torch version to 2.10 for simplicity

* fix: namespace cache suffix by torch and cuda version

* fix: add tashes to cache-suffix

+++ src/vision3d/ops/csrc/iou_box3d/iou_utils.h
@@ -321,7 +321,7 @@ inline float BoxVolume(
   float box_vol = 0.0;
   // Iterate through each triange, calculate the area of the
   // tetrahedron formed with the box_center and sum them
-  for (int t = 0; t < box_tris.size(); ++t) {
+  for (size_t t = 0; t < box_tris.size(); ++t) {
     // Subtract the center:
     const vec3<float> v0 = box_tris[t][0] - box_center;
     const vec3<float> v1 = box_tris[t][1] - box_center;
@@ -722,11 +722,11 @@ inline face_verts BoxIntersections(
     // Iterate through triangles in tris
     // Create intermediate vector to store the updated tris
     face_verts tri_verts_updated;
-    for (int t = 0; t < out_tris.size(); ++t) {
+    for (size_t t = 0; t < out_tris.size(); ++t) {
       // Clip tri by plane
       const face_verts tri_updated = ClipTriByPlane(planes[p], out_tris[t], n2);
       // Add to the tri_verts_updated output if not empty
-      for (int v = 0; v < tri_updated.size(); ++v) {
+      for (size_t v = 0; v < tri_updated.size(); ++v) {
         tri_verts_updated.push_back(tri_updated[v]);
       }
     }
